fix(adapter): default adapter dtype to float32 — bf16 weights stall under plain AdamW
torch.optim.AdamW keeps its state in the param dtype and applies updates
in-place: with bf16 adapter weights (old default = model dtype), any update
below ~0.2% of a weight's magnitude rounds to zero, so training silently
stalls — flat train/loss with a healthy grad norm. Measured on a synthetic
persistent-gradient run: at lr 1e-4 bf16 params move ~18% of what fp32
moves over 400 steps. kohya, ai-toolkit and PEFT all force adapter weights
to fp32 for exactly this reason (ai-toolkit: network.force_to(dtype=fp32)
before building the optimizer).

Adapter weights now default to float32 for lora/lokr/lycoris regardless of
model dtype; autocast still runs the forward matmuls in the model dtype so
speed is unchanged. An explicit 16-bit adapter.dtype still works but prints
a NOTE pointing at Kahan/stochastic-rounding optimizers (adamw8bitkahan,
adamw_optimi), which are the only safe pairing.

+++ tests/test_lycoris_sdxl_network.py
@@ -308,7 +308,9 @@ def test_defaults_fill_per_algo_and_reject_alpha():
     assert adapter["dora_wd"] is False
     assert adapter["wd_on_output"] is True
     assert adapter["train_conv"] is False
-    assert adapter["dtype"] == torch.bfloat16
+    # float32 regardless of model dtype: bf16 adapter weights lose sub-0.2% AdamW updates
+    # to rounding (flat loss, healthy grad norm).
+    assert adapter["dtype"] == torch.float32
 
     with pytest.raises(ConfigValidationError, match="alpha"):
         set_config_defaults(_base_config({"type": "lycoris_locon", "rank": 4, "alpha": 2}))
